requirements.txt
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
boto3==1.17.111
botocore==1.20.111
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
dateutils==0.6.12
distro-info===0.18ubuntu0.18.04.1
Django==3.2.4
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-storages==1.11.1
djangorestframework==3.12.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mysqlclient==2.0.3
netifaces==0.10.4
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
pymemcache==3.0.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-memcached==1.59
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.4.2
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.4.1
ssh-import-id==5.7
swapper==1.1.2.post1
systemd-python==234
Twisted==17.9.0
typing-extensions==3.10.0.0
ufw==0.36
urllib3==1.26.6
wrapt==1.12.1
zope.interface==4.3.2
